añadido el tipo float al precio del producto y los atributos beginDate y endDate a prmociones

+++ DSS/database/migrations/2021_03_02_153020_create_products_table.php
@@ -17,8 +17,8 @@ class CreateProductsTable extends Migration
             $table->bigIncrements('id');
             $table->timestamps();
             $table->string('name');
-            $table->integer('price');
-            $table->integer('promotionPrice');
+            $table->float('price');
+            $table->float('promotionPrice');
             $table->string('description');
             $table->integer('stock');
             $table->string('color');
